refactor mobile push to avoid a type of error
