+++ go/bind/keybase.go
@@ -433,8 +433,8 @@ func BackgroundSync() {
 	<-doneCh
 }
 
-func HandleBackgroundNotification(strConvID, body string, intMembersType int, displayPlaintext bool, intMessageID int,
-	pushID string, badgeCount, unixTime int, soundName string, pusher PushNotifier) (err error) {
+func HandleBackgroundNotification(strConvID, body string, intMembersType int, displayPlaintext bool,
+	intMessageID int, pushID string, badgeCount, unixTime int, soundName string, pusher PushNotifier) (err error) {
 	if err := waitForInit(5 * time.Second); err != nil {
 		return nil
 	}
@@ -447,15 +447,33 @@ func HandleBackgroundNotification(strConvID, body string, intMembersType int, di
 		func() error { return err })()
 	defer func() { err = flattenError(err) }()
 
-	msg, err := unboxNotification(ctx, strConvID, body, intMembersType)
+	// Unbox
+	if !kbCtx.ActiveDevice.HaveKeys() {
+		return libkb.LoginRequiredError{}
+	}
+	mp := chat.NewMobilePush(gc)
+	uid := gregor1.UID(kbCtx.Env.GetUID().ToBytes())
+	bConvID, err := hex.DecodeString(strConvID)
 	if err != nil {
+		kbCtx.Log.CDebugf(ctx, "HandleBackgroundNotification: invalid convID: %s msg: %s", strConvID, err)
 		return err
 	}
-
+	convID := chat1.ConversationID(bConvID)
+	membersType := chat1.ConversationMembersType(intMembersType)
+	msgUnboxed, err := mp.UnboxPushNotification(ctx, uid, convID, membersType, body)
+	if err != nil {
+		kbCtx.Log.CDebugf(ctx, "unboxNotification: failed to unbox: %s", err)
+		return err
+	}
 	if !displayPlaintext {
 		return nil
 	}
 
+	// Send notification
+	msg, err := mp.FormatPushText(ctx, uid, convID, membersType, msgUnboxed)
+	if err != nil {
+		return err
+	}
 	age := time.Since(time.Unix(int64(unixTime), 0))
 	if age >= 15*time.Second {
 		kbCtx.Log.CDebugf(ctx, "HandleBackgroundNotification: stale notification: %v", age)
@@ -465,29 +483,8 @@ func HandleBackgroundNotification(strConvID, body string, intMembersType int, di
 	id := fmt.Sprintf("%s:%d", strConvID, intMessageID)
 	pusher.LocalNotification(id, msg, badgeCount, soundName, strConvID, "chat.newmessage")
 	// Hit the remote server to let it know we succeeded in showing something useful
-	kbCtx.ChatHelper.AckMobileNotificationSuccess(ctx, []string{pushID})
+	mp.AckNotificationSuccess(ctx, []string{pushID})
 	return nil
-}
-
-func unboxNotification(ctx context.Context, strConvID, body string, intMembersType int) (msg string, err error) {
-	if !kbCtx.ActiveDevice.HaveKeys() {
-		return "", libkb.LoginRequiredError{}
-	}
-	uid := gregor1.UID(kbCtx.Env.GetUID().ToBytes())
-	bConvID, err := hex.DecodeString(strConvID)
-	if err != nil {
-		kbCtx.Log.CDebugf(ctx, "unboxNotification: invalid convID: %s msg: %s", strConvID,
-			err)
-		return "", err
-	}
-	convID := chat1.ConversationID(bConvID)
-	membersType := chat1.ConversationMembersType(intMembersType)
-	msg, err = kbCtx.ChatHelper.UnboxMobilePushNotification(ctx, uid, convID, membersType, body)
-	if err != nil {
-		kbCtx.Log.CDebugf(ctx, "unboxNotification: failed to unbox: %s", err)
-		return "", err
-	}
-	return msg, nil
 }
 
 func pushPendingMessageFailure(convID chat1.ConversationID, pusher PushNotifier) {
