+++ go/libkb/interfaces.go
@@ -801,9 +801,6 @@ type ChatHelper interface {
 	GetMessages(ctx context.Context, uid gregor1.UID, convID chat1.ConversationID,
 		msgIDs []chat1.MessageID, resolveSupersedes bool) ([]chat1.MessageUnboxed, error)
 	UpgradeKBFSToImpteam(ctx context.Context, tlfName string, tlfID chat1.TLFID, public bool) error
-	UnboxMobilePushNotification(ctx context.Context, uid gregor1.UID,
-		convID chat1.ConversationID, membersType chat1.ConversationMembersType, payload string) (string, error)
-	AckMobileNotificationSuccess(ctx context.Context, pushIDs []string)
 }
 
 // Resolver resolves human-readable usernames (joe) and user asssertions (joe+joe@github)
